workshop2/pyproject.toml
[project]
name = "workshop2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "ipywidgets>=8.1.7",
    "logfire[requests]>=4.2.0",
    "markdown>=3.8.2",
    "minsearch>=0.0.4",
    "notebook>=7.4.4",
    "openai>=1.97.0",
    "pydantic-ai>=0.5.0",
    "pydantic-ai-slim[mcp]>=0.5.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
